Refactoring: move dataset loading to own package
